Moved Show more/less button with a number of hubs prop to  MiniHubPreviews
