+++ frontend/src/components/account/AccountPage.js
@@ -11,7 +11,7 @@ import {
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import PlaceIcon from "@material-ui/icons/Place";
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
 import Linkify from "react-linkify";
 import Cookies from "universal-cookie";
 
@@ -25,8 +25,6 @@ import ProfileBadge from "../profile/ProfileBadge";
 import SocialMediaShareButton from "../shareContent/SocialMediaShareButton";
 import UserContext from "../context/UserContext";
 import EditSharpIcon from "@material-ui/icons/EditSharp";
-import ExpandLessIcon from "@material-ui/icons/ExpandLess";
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 
 const useStyles = makeStyles((theme) => ({
   avatarContainer: {
@@ -156,12 +154,6 @@ const useStyles = makeStyles((theme) => ({
   miniOrgPreview: {
     display: "flex",
   },
-  
-  hubButtonsContainer: {
-    display: "flex",
-    justifyContent: "center",
-    marginLeft: "-35%",
-  },
 }));
 
 //Generic component to display personal profiles or organization profiles
@@ -200,11 +192,7 @@ export default function AccountPage({
       {text}
     </Link>
   );
-  const [showMoreSectors, setShowMoreSectors] = useState(false);
 
-  const handleShowMoreSectors = () => {
-    setShowMoreSectors(!showMoreSectors);
-  };
   const displayAccountInfo = (info) =>
     Object.keys(info)
       .sort((a, b) => {
@@ -260,34 +248,19 @@ export default function AccountPage({
               </div>
             );
           } else if (i.type === "hubs") {
-            if (i.value.length <= 2) return <MiniHubPreviews hubs={i.value} />;
-            else {
-              const hubsToDisplay = showMoreSectors ? i.value : i.value.slice(0, 2);
-              return (
-                <>
-                  <div className={classes.subtitle}>
-                    {" "}
-                    {organizationTexts.organization_is_active_in_these_sectors}{" "}
-                  </div>
-                    <MiniHubPreviews hubs={hubsToDisplay} />
-                    <div className={classes.hubButtonsContainer}>
-                      <Button onClick={handleShowMoreSectors}>
-                        {showMoreSectors ? (
-                          <>
-                            <ExpandLessIcon />
-                            {texts.show_less}
-                          </>
-                        ) : (
-                          <>
-                            <ExpandMoreIcon />
-                            {texts.show_more}
-                          </>
-                        )}
-                      </Button>
-                    </div> 
-                </>
-              );
-            }
+            return (
+              <>
+                <div className={classes.subtitle}>
+                  {" "}
+                  {organizationTexts.organization_is_active_in_these_sectors}{" "}
+                </div>
+                {i.value.length <= 2 ? (
+                  <MiniHubPreviews hubs={i.value} texts={texts} />
+                ) : (
+                  <MiniHubPreviews hubs={i.value} maxHubsToShow={2} texts={texts} />
+                )}
+              </>
+            );
           } else if (i.type === "select" && value) {
             const textValue = i.options ? i.options.find((o) => o?.key === value).name : value;
             return (
