+++ frontend/src/components/hub/MiniHubPreviews.js
@@ -1,6 +1,8 @@
-import { Grid, makeStyles } from "@material-ui/core";
-import React from "react";
+import { Grid, makeStyles, Button } from "@material-ui/core";
+import React, { useState } from "react";
 import MiniHubPreview from "./MiniHubPreview";
+import ExpandLessIcon from "@material-ui/icons/ExpandLess";
+import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 
 const useStyles = makeStyles((theme) => ({
   reset: {
@@ -18,6 +20,11 @@ const useStyles = makeStyles((theme) => ({
   lastItem: {
     marginRight: theme.spacing(-1),
   },
+  hubButtonsContainer: {
+    display: "flex",
+    justifyContent: "center",
+    marginLeft: "-35%",
+  },
 }));
 
 export default function MiniHubPreviews({
@@ -27,31 +34,57 @@ export default function MiniHubPreviews({
   editMode,
   onSelectNewHub,
   onClickRemoveHub,
+  maxHubsToShow,
+  texts,
 }) {
   const classes = useStyles();
+  const [showMoreSectors, setShowMoreSectors] = useState(false);
+  const handleShowMoreSectors = () => {
+    setShowMoreSectors(!showMoreSectors);
+  };
+  const hubsToDisplay = showMoreSectors ? hubs : hubs.slice(0, maxHubsToShow);
+
   return (
-    <Grid container component="ul" spacing={2} className={`${classes.reset} ${classes.root}`}>
-      {hubs
-        .filter((hub) => hub.hub_type !== "location hub")
-        .map((hub, index) => (
+    <>
+      <Grid container component="ul" spacing={2} className={`${classes.reset} ${classes.root}`}>
+        {hubsToDisplay
+          .filter((hub) => hub.hub_type !== "location hub")
+          .map((hub, index) => (
+            <GridItem
+              hub={hub}
+              key={hub.url_slug}
+              editMode={editMode}
+              isFirstItem={index === 0}
+              onClickRemoveHub={onClickRemoveHub}
+            />
+          ))}
+        {allowCreate && (
           <GridItem
-            hub={hub}
-            key={hub.url_slug}
-            editMode={editMode}
-            isFirstItem={index === 0}
-            onClickRemoveHub={onClickRemoveHub}
+            createMode
+            isFirstItem={hubs.length === 0}
+            isLastItem
+            hubsToSelectFrom={allHubs}
+            onSelectNewHub={onSelectNewHub}
           />
-        ))}
-      {allowCreate && (
-        <GridItem
-          createMode
-          isFirstItem={hubs.length === 0}
-          isLastItem
-          hubsToSelectFrom={allHubs}
-          onSelectNewHub={onSelectNewHub}
-        />
-      )}
-    </Grid>
+        )}
+      </Grid>
+      {!editMode && ( <div className={classes.hubButtonsContainer}>
+        <Button onClick={handleShowMoreSectors}>
+          {showMoreSectors ? (
+            <>
+              <ExpandLessIcon />
+              {texts.show_less}
+            </>
+          ) : (
+            <>
+              <ExpandMoreIcon />
+              {texts.show_more}
+            </>
+          )}
+        </Button>
+      </div>)}
+     
+    </>
   );
 }
